+++ dart_exporter/example_package/lib/src/exports.dart_exporter.dart
@@ -1,6 +1,0 @@
-// example_package
-// AUTO GENERATE FILE, DONT MODIFY!!
-// ignored: 
-
-export 'package:example_package/src/khoa1_copy.dart' show Z2;
-export 'package:example_package/src/test.dart' show Z,ZExtension,E;