+++ dart_exporter/example_package/lib/src/khoa1.dart
@@ -1,7 +1,0 @@
-part of 'test.dart';
-
-@Export()
-class Z {}
-
-@Export()
-extension ZExtension on Z {}