+++ dart_exporter/example_package/lib/src/khoa1_copy.dart
@@ -1,4 +1,0 @@
-import 'package:dart_exporter_annotation/dart_exporter_annotation.dart';
-
-@Export()
-class Z2 {}