+++ dart_exporter/example_package/lib/src/test.dart
@@ -1,14 +1,11 @@
 import 'package:dart_exporter_annotation/dart_exporter_annotation.dart';
 
-part 'khoa1.dart';
-part 'test2.dart';
-
 class A {}
 
-class B {}
+@Export()
+class AExport {}
 
-class _A {}
-
-class C {}
-
-class D {}
+@Export(
+  forceExportItems: ['test_'],
+)
+class AExportPremium {}