+++ dart_exporter/example_package/lib/src/test2.dart
@@ -1,4 +1,0 @@
-part of 'test.dart';
-
-@Export()
-class E {}