+++ dart_exporter_annotation/lib/dart_exporter_annotation.dart
@@ -11,5 +11,8 @@ class DoNotExport {
 }
 
 class Export {
-  const Export();
+  final List<String> forceExportItems;
+  const Export({
+    this.forceExportItems = const [],
+  });
 }