+++ dart_exporter/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 3.0.2
+- update dart exporter annotation with ```forceExportItems```
+
 ## 3.0.1
 - update pubspecs